# backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: Idento API
  version: "1.0"
  description: Idento event check-in platform REST API.
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      description: >
        Standard handler error: {"error": "..."} with an optional machine code
        (e.g. code=tenant_suspended on 403 from the tenant gate). Closed shape
        (additionalProperties: false) so it never ambiguously matches the
        richer LimitExceededError body in a oneOf.
      properties:
        error: { type: string }
        code: { type: string }
      required: [error]
      additionalProperties: false
...
    LimitExceededError:
      type: object
      description: >
        middleware.CheckLimits 403 body when a tenant resource limit
        (e.g. users) is at or over its plan max.
      properties:
        error: { type: string }
        current: { type: integer }
        max: { type: integer, description: "-1 means unlimited" }
        upgrade_required: { type: boolean }
        limit_type: { type: string }
      required: [error, current, max, upgrade_required, limit_type]
      additionalProperties: false
...
    Attendee:
      type: object
      properties:
        id: { type: string, format: uuid }
        event_id: { type: string, format: uuid }
        first_name: { type: string }
        last_name: { type: string }
        email: { type: string }
        company: { type: string }
        position: { type: string }
        code: { type: string }
        checkin_status: { type: boolean }
        checked_in_at: { type: string, format: date-time, nullable: true }
        checked_in_by: { type: string, format: uuid, nullable: true }
        checked_in_by_email:
          type: string
          nullable: true
          description: Email of the user who checked the attendee in (set once, on the first check-in transition).
        checked_in_device_number: { type: integer, nullable: true }
        checked_in_point_name: { type: string, nullable: true }
        printed_count: { type: integer }
        blocked: { type: boolean }
        block_reason: { type: string, nullable: true }
        packet_delivered: { type: boolean }
        registered_at: { type: string, format: date-time, nullable: true }
        registration_zone_id: { type: string, format: uuid, nullable: true }
        custom_fields: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: Soft-delete marker set by DELETE /api/attendees/{id}; the row is never hard-deleted.
      required:
        [
          id,
          event_id,
          first_name,
          last_name,
          email,
          company,
          position,
          code,
          checkin_status,
          printed_count,
          blocked,
          packet_delivered,
          created_at,
          updated_at,
        ]
    AttendeeListPage:
      type: object
      description: >
        Envelope returned by GET /api/events/{event_id}/attendees when the
        `page` or `per_page` query param is present. `total` is the count of
        attendees matching the active filters (code/search/zone/status)
        before paging is applied, not the length of `attendees`.
      properties:
        attendees:
          type: array
          items: { $ref: "#/components/schemas/Attendee" }
        total: { type: integer, description: Total matching attendees across all pages. }
        page: { type: integer }
        per_page: { type: integer }
      required: [attendees, total, page, per_page]
      additionalProperties: false
...
paths:
...
  /api/events/{event_id}/attendees:
    get:
      operationId: getAttendees
      summary: >
        Attendees for an event, optionally filtered by an exact code match
        (mobile QR/barcode scan lookup) and/or a search substring across
        name/email/code (mobile attendee search). Omitting both query
        params returns the full event attendee list.
      description: >
        Response shape is a discriminated-by-param-presence oneOf: when
        NEITHER `page` nor `per_page` is present, the response is the
        legacy bare array of Attendee (back-compat for mobile clients,
        unchanged). When EITHER `page` or `per_page` is present, the
        response is an AttendeeListPage envelope instead, and `zone` /
        `status` additionally narrow the result set. `page` defaults to 1
        and `per_page` defaults to 50 when only one of the pair is given.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: code
          in: query
          schema: { type: string }
          description: Exact match against Attendee.code.
        - name: search
          in: query
          schema: { type: string }
          description: Substring match across name/email/code.
        - name: page
          in: query
          schema: { type: integer, minimum: 1 }
          description: >
            1-indexed page number. Presence of this param (or per_page)
            switches the response to the AttendeeListPage envelope.
            Defaults to 1 if only per_page is given. Must be >= 1.
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 200 }
          description: >
            Page size. Presence of this param (or page) switches the
            response to the AttendeeListPage envelope. Defaults to 50 if
            only page is given. Must be between 1 and 200 inclusive.
        - name: zone
          in: query
          schema: { type: string, format: uuid }
          description: >
            Event zone UUID. Only applied in the paginated envelope mode
            (page or per_page present); narrows to attendees with an
            explicit attendee_zone_access row for that zone with
            allowed=true.
        - name: status
          in: query
          schema: { type: string, enum: [checked_in, not_checked_in] }
          description: >
            Only applied in the paginated envelope mode (page or per_page
            present); filters on Attendee.checkin_status.
      responses:
        "200":
          description: >
            Attendees for the event. A bare array when neither page nor
            per_page is present (legacy shape, unfiltered by zone/status);
            an AttendeeListPage envelope otherwise.
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: "#/components/schemas/Attendee" }
                  - { $ref: "#/components/schemas/AttendeeListPage" }
        "400":
          description: >
            event_id is not a UUID, or (paginated mode only) page < 1,
            per_page outside 1..200, zone is not a UUID, or status is
            neither checked_in nor not_checked_in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: tenant_suspended from the tenant gate.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: >
            Event does not exist, or belongs to a different tenant
            (requireEventOwnership masks "foreign" as "missing").
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: >
            Store failure resolving event ownership ("Internal error"), or
            fetching attendees ("Failed to fetch attendees") — both render
            as Error via c.JSON.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      operationId: createAttendee
      summary: Create an attendee for an event
      security: [{ bearerAuth: [] }]
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: >
                No field is required by the handler — an empty body creates
                an attendee with every field blank. code defaults to a
                generated UUID string when omitted or empty.
              properties:
                first_name: { type: string }
                last_name: { type: string }
                email: { type: string }
                company: { type: string }
                position: { type: string }
                code: { type: string }
                custom_fields: { type: object, additionalProperties: true }
      responses:
        "201":
          description: >
            Created attendee. Usage is logged best-effort afterward (a
            LogUsage failure is logged server-side but does not fail the
            request), same pattern as POST /api/events.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attendee" }
        "400":
          description: >
            event_id is not a UUID — checked identically by
            middleware.CheckAttendeeLimits before the handler runs, so the
            handler's own duplicate check on this route is unreachable — or
            the request body is malformed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: >
            tenant_suspended from the tenant gate (Error), or
            middleware.CheckAttendeeLimits rejected the request because the
            event is at/over its attendees_per_event plan limit
            (LimitExceededError). For this single-create route (adding
            always =1), CheckAttendeeLimits happens to build the exact same
            body shape as middleware.CheckLimits even though it's a
            separate function — contrast BulkLimitExceededError on the
            /bulk route, where "adding" (the batch size) is a real extra
            field.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/Error" }
                  - { $ref: "#/components/schemas/LimitExceededError" }
        "404":
          description: >
            Event does not exist, or belongs to a different tenant
            (requireEventOwnership, checked by the handler after the
            middleware passes).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: >
            middleware.CheckAttendeeLimits' own store failure checking the
            limit ("Failed to check attendee limit" — unlike
            middleware.CheckLimits, which folds a store error into the same
            403 as a real limit breach, CheckAttendeeLimits surfaces a
            store failure honestly as a 500), a store failure resolving
            event ownership ("Internal error"), or a failure persisting the
            attendee ("Failed to create attendee"). All render as Error via
            c.JSON.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...